requirements.txt
requests[socks]
bs4
pandas
fpdf2
